@@ -1,4 +1,4 @@
-module github.com/cheggaaa/pb/v3
+module github.com/dmitryk-dk/pb/v3
 
 require (
 	github.com/VividCortex/ewma v1.1.1
@@ -6,8 +6,11 @@ require (
 	github.com/mattn/go-colorable v0.1.8
 	github.com/mattn/go-isatty v0.0.12
 	github.com/mattn/go-runewidth v0.0.12
-	github.com/rivo/uniseg v0.2.0 // indirect
-	golang.org/x/sys v0.0.0-20210403161142-5e06dd20ab57 // indirect
 )
 
-go 1.12
+require (
+	github.com/rivo/uniseg v0.2.0 // indirect
+	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
+)
+
+go 1.18
