@@ -1,5 +1,4 @@
-// +build darwin freebsd netbsd openbsd dragonfly
-// +build !appengine
+//go:build (darwin || freebsd || netbsd || openbsd || dragonfly) && !appengine
 
 package termutil
 
