@@ -1,4 +1,4 @@
-// +build windows
+//go:build windows
 
 package termutil
 
@@ -111,7 +111,7 @@ func termWidthTPut() (width int, err error) {
 	return strconv.Atoi(string(res))
 }
 
-func getCursorPos() (pos coordinates, err error) {
+func GetCursorPos() (pos coordinates, err error) {
 	var info consoleScreenBufferInfo
 	_, _, e := syscall.Syscall(procGetConsoleScreenBufferInfo.Addr(), 2, uintptr(syscall.Stdout), uintptr(unsafe.Pointer(&info)), 0)
 	if e != 0 {
@@ -120,7 +120,7 @@ func getCursorPos() (pos coordinates, err error) {
 	return info.dwCursorPosition, nil
 }
 
-func setCursorPos(pos coordinates) error {
+func SetCursorPos(pos coordinates) error {
 	_, _, e := syscall.Syscall(setConsoleCursorPosition.Addr(), 2, uintptr(syscall.Stdout), uintptr(uint32(uint16(pos.Y))<<16|uint32(uint16(pos.X))), 0)
 	if e != 0 {
 		return error(e)
